// src/types/messages/pages/login.ts
enum LoginPageTranslations {
  LOGIN = 'pages.login.login',
  USERNAME_LABEL = 'pages.login.usernameLabel',
  USERNAME_PLACEHOLDER = 'pages.login.usernamePlaceholder',
  USERNAME_REQUIRED_ERROR = 'pages.login.usernameRequiredError',
  PASSWORD_LABEL = 'pages.login.passwordLabel',
  PASSWORD_PLACEHOLDER = 'pages.login.passwordPlaceholder',
  PASSWORD_REQUIRED_ERROR = 'pages.login.passwordRequiredError',
  LOGIN_SUCCESS = 'pages.login.loginSuccess',
  DONT_HAVE_ACCOUNT = 'pages.login.dontHaveAccount',
  REGISTER_NOW = 'pages.login.registerNow',
}

export default LoginPageTranslations;